Move method: TrySetKubectlVersion
